% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{anova}
\alias{anova}
\title{方差分析}
\usage{
anova(data, y_colname, x_colname)
}
\arguments{
\item{data}{进行分析的数据}

\item{y_colname}{定量数据列名}

\item{x_colname}{定类数据列名}
}
\value{
df，平方和，均方，f值，p值
}
\description{
方差分析
}
