% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_analysis}
\alias{cross_analysis}
\title{交叉分析}
\usage{
cross_analysis(data, handled = FALSE, method = "chisq")
}
\arguments{
\item{data}{分析对象数据}

\item{handled}{对象数据是否分类并统计各类数目}

\item{method}{要使用的方法，只能是"chisq"和"fisher"二者之一}
}
\value{
分析结果
}
\description{
交叉分析
}
